Run Kotlin in process during integration tests
We seem to be running into sporadic errors of the form:
 Error unmarshaling return header; nested exception is:
  java.io.EOFException
  at sun.rmi.transport.StreamRemoteCall.executeCall(StreamRemoteCall.java:229)
  at sun.rmi.server.UnicastRef.invoke(UnicastRef.java:162)
  at java.rmi.server.RemoteObjectInvocationHandler.invokeRemoteMethod(RemoteObjectInvocationHandler.java:227)
  at java.rmi.server.RemoteObjectInvocationHandler.invoke(RemoteObjectInvocationHandler.java:179)
  at com.sun.proxy.$Proxy117.clearJarCache(Unknown Source)
  at org.jetbrains.kotlin.compilerRunner.GradleCompilerRunner.compileWithDaemon(GradleKotlinCompilerRunner.kt:216)
... both in our post-submit and on go/ab, and one theory is that the
Kotlin compiler daemon is running out of memory and unresponsive/dead
(it is running with -Xmx1024m, not too much).

If we run the compiler in process, it has access to a larger heap (4Gb)
and hopefully we get more information in the logs if it crashes.

Test: N/A
Change-Id: I3795b0b95b0bb310e4a82fb2317849de2361b16b