Consider device characteristics when invoking simpleperf
When invoking simpleperf, we need to take into account the features
supported by the target device. Emulators, for instance, don't support
cpu-cycles hardware events, so we need to use cpu-clock events instead.
We also need to check if the device supports dwarf-based call graphs or
use frame pointers otherwise. Finally, we should check if the device
supports offcpu tracing and pass the corresponding flag if it does.

This CL also refactors Simpleperf::Record logic to break it into two
methods: one to generate the `simpleperf record` command, passing all
the flags and arguments; and another one to actually invoke simpleperf
with such command. This new logic makes the code unit testable and is
covered by the newly added tests.

Fixes: 68013630
Bug: 37294023

Test: simpleperf_test.cc

Change-Id: I5649ff8fa336a800c748148dd31a403e525706c5