Add warnings on improper usage of generatePureSplits
Users should now see a warning when they enable language splits without
setting generatePureSplits, in both AS and command line builds, because
language splits only work with Configuration APKS.

Also moved a warning about Pure Splits from BaseExtension to BasePlugin,
and set it to only surface when the module is not using the Feature
plugin.

In a similar vein, added a warning for users who attempt to use
non-pure splits in a feature module, and set the MultiOutputPolicy to
Splits instead of Multi-APK since multi-APK does nothing while building
a feature module.

Bug: b/67869307
Test: Verified warning content in PureSplits integration tests.
Change-Id: I68d096589d523ca41e60380b50dc4f20e51e2b40