Add the concept of UploadStrategy to BenchmarkRecorder.
This will allow us to run benchmarks N times and take the smallest in
a non-disruptive, low-effort way.

There's a small problem: this uses the overall build time as the
heuristic. We graph data more granularly than this. This introduces
a potential mismatch between the way we record data and the way we
decide to upload it. I decided, though, that the cost of doing this
correctly is outweighed by the increased value in doing it well enough.

Test: presubmit
Change-Id: I6da37c3d95731a2d079431e212f717a0917fb590