Skeleton of the structure plugin.
the structure plugin is a debug plugin that can be applied
to a bunch of projects to gather structure information.

This allows external dev to run it, give the output to us
so that we can repro a similar project (minus code and resources).

There is small change to BasePlugin to make it more easy to use.
- BasePlugin now instantiate Plugin<Project>. No reason it didn't before
- getExtension is now public. This is actually a good api to have instead
  of haaving to do get("Android") on the extensio container and then casting.

Change-Id: I734ea3b5a60dadee16751abe7c491f7fae362cdb
Test: N/A