Sessions-related proto update and cleanup
- Added device_id which uniquely identifies a Device
based on its serial and boot_id. The device_id is generated
from a SHA from Studio-side based on those two properties.
As a result, some device-specific rpc requests have been
changed to take in only the device_id instead of an entire
Device proto object.

- Updated the Session proto to reference the device_id
of a Device, removing the need to reference both the
serial and boot_id information

- Changed session_id generation to take into account the
device_id

Bug: 67508221
Test: Updated test cases

Change-Id: I7681d83a2bf38d6244f575f5e4e301f84238a3eb