Gradle 4.4 rc3 + bazel clean up
This involves manually adding some artifacts to the
offline repo as Gradle resolve dependencies slightly differently
from Bazel.

This also cleans up the different repo we had in different location
to put them all in //tools/base/build-system and have them extend
each other so that it's clear what's going on.

Also update all rules and code to deal with the removal
of the automated _repo suffix added to all repo.
It now only add the .zip extension.

Change-Id: Ibf67883c7c2996ef29a5e97219317c83da8f445b
Test: existing tests.