Implement eviction policy for Android Gradle build cache.
At regular intervals, we will delete old cache entries from the Android
Gradle build cache. Cache eviction is triggered at the beginning of the
build when the specified interval has passed since the last cache
eviction.

The cache eviction interval is currently set to 1 day.

The cache entry life time (number of days that cache entries are kept in
the build cache) is currently set to 30 days.

We can allow the users to customize both the cache eviction interval and
the cache entry life time later if there is a need for it.

Bug: 69562033
Test: Updated FileCacheTest, BuildCacheUtilsTest
Change-Id: I1f4929c4f8e92de8f0f02800015a38fa90efc400