Fix long overflow in ActdProfileUploader test.
This was a fun one. We were generating random durations between 1 and
Long.MAX_LONG. At some point in the process we end up summing durations
for samples that share a build ID and task type, and some of these
naturally overflow. If _all_ of them overflow and end up negative, we
filter them out because they're lower than our threshold of 50ms for
tasks to reduce the effect of noisy, quick tasks in our data.

Test: ActdProfileUploaderTest
Change-Id: I1b13c7cbe8c0959ad05eff3cb9e20957fdd73e87