[AIA] Generate aia protos in (pre) build time
This CL replaces a dependency on a pre-generated jar with compiled proto
classes for instant app and replaces with dependencies on jars generated
in build time.

When in tools/idea, protos are compiled and then jarjared. Library
dependency in iml points to expected output for making IDE happy.

When in tools/base, protos are just compiled directly with proto3.
Library dependency in iml also points to expected generated folder.

This CL also modifies domain of repackage to avoid conflicts.

Test: all existing tests and build
Change-Id: I8b19a1efcb1df0823513793b888545b4651bd2f6