Deprecated 'testApi' and 'androidTestApi" in the old DSL
Added deprecation warnings for all "test___Api" and "androidTest___Api"
configurations, which required refactoring:

Moved the work of creating the Configurations for each SourceSet out of
a whenObjectAdded() block on the sourceSets DSL object and into a new
SourceSetManager class. This code is still paired with SourceSet
creation as long as we use the maybeCreateAndConfigure() method instead
of calling maybeCreate() directly on the backing container.

If a SourceSet is found at variant creation time for which the correct
configurations have not been created, a SyncIssue will be emitted. This
is most likely to be the result of a typo in a sourceSets block.

Bug: b/69710020
Test: Added MistypedSourceSetTest
Change-Id: Ic3c0d0e14fbeccd54bc53efad5d62390b4682519