Add manifest merger error if no package in library manifest
Currently, there's a warning if the library package is not
declared. This CL changes this warning to an error and also
adds an error if it's declared as an empty string.

This CL also adds package="${packageName}" to many of the
template manifests.

This CL also tweaks ManifestMerger2Test to check the expected
errors/warnings for a test case even if the runtime test merging report
doesn't contain any errors or warnings (and deletes some obsolete
expected warnings from a couple test cases).

Bug: 71857930
Test: added test case to ManifestMerger2Test
Change-Id: Ie67810021f69672cae61bfe09a42882172d83f37