Resource shrinker processing dex files
This CL adds support to resource shrinker to
analyze dex files. This is necessary as R8 produces
dex files, and reachability analysis needs to
be performed on dex files now.

Dex file visitor has been added, which analyzes
all classes in a dex files. All fields, methods,
and annotations are analyzed in the same way
as when processing bytecode. Resource analyzer
will run dex or class visitor based on the
file extension, so for ResourceUsageAnalyzer
users, this change is transparent.

ShrinkResourceTransform has been changed to
accept DEX and CLASSES as input types, as it
can run after R8Transform, or ProguardTransform.

BUG: 72370175
Test: running existing integration tests with R8,
ResourceUsageAnalyzerTest
Change-Id: I73e55816a65970d2b6aefaf8655a3f468998bea8