Disable tests on foundry that are consistently failing on foundry bot.
This is an example of the tests that are most consistently failing:
go/upsalite-prod/3f7575dc-0c05-4245-9e8e-c0bac12a7a8b/targets
There are other tests that fail too, but these are the ones that fail the most.
Let's disable these and I'll disable the other ones as they show up. This should
make the foundry bot have green builds.

Bug: 73948555
Test: No PSQ tests yet, need to see results on post-submit bot
Change-Id: I66b7c25b267530e4a9e8706ae69d77924ef40476